RS_Darkness: spare actors from the world darkness
SetDarknessActors(exempt) is back, and with it FLevelLocals::DarkActorExempt.

The darkness pass takes the whole scene down together, which takes the monsters
with it -- correct for a wall and wrong for the thing walking towards you, which
simply stops existing. A room dark enough to be worth carrying a light into was
a room you could not see anything coming in.

uDarknessExempt goes from int to float, in all three backends, because the
useful settings are between the ends: 0 takes all of the darkening, 1 takes
none, and an actor sits somewhere in the middle so a blacked-out room still has
things visible moving in it. The shader mixes toward 1 rather than testing == 0.

Both draw paths take it. Fullbright things are exempt outright; everything else
gets what the level asked for. The psprite path gets the same value as the
sprite path deliberately -- the VR hands can be drawn either way, toggled at
runtime, and the same hands darkening differently depending on which mode they
were in is invisible until you switch and then inexplicable.

Defaults to 0, which is the stock behaviour, so nothing changes until the mod
asks. First of the four lighting mods being restored one at a time.

+++ src/common/rendering/gl/gl_shader.cpp
@@ -453,7 +453,7 @@ bool FShader::Load(const char * name, const char * vert_prog_lump, const char * 
 		uniform vec4 uFlatGlowFar;
 		uniform int uFlatGlowFalloff;
 		uniform int uFlatGlowIsCeiling;
-		uniform int uDarknessExempt;
+		uniform float uDarknessExempt;
 		uniform int uFlatGlowLineCount;
 		uniform vec4 uFlatGlowLines[64];
